assessment/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
httpx>=0.27.0
openpyxl>=3.1.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-multipart>=0.0.6
PyJWT>=2.8.0
sqlalchemy[asyncio]>=2.0.0
aiosqlite>=0.19.0
asyncpg>=0.29.0
